+++ src/Infrastructure/NotificationEvent/NotifyFailedException.php
@@ -1,0 +1,10 @@
+<?php
+
+namespace ArtARTs36\MergeRequestLinter\Infrastructure\NotificationEvent;
+
+use ArtARTs36\MergeRequestLinter\Shared\Exceptions\MergeRequestLinterException;
+
+final class NotifyFailedException extends MergeRequestLinterException
+{
+    //
+}