+++ src/Infrastructure/Notifications/Notifier/MessageCreator.php
@@ -4,6 +4,7 @@ namespace ArtARTs36\MergeRequestLinter\Infrastructure\Notifications\Notifier;
 
 use ArtARTs36\MergeRequestLinter\Domain\Notifications\Message;
 use ArtARTs36\MergeRequestLinter\Infrastructure\Contracts\Text\TextRenderer;
+use ArtARTs36\MergeRequestLinter\Infrastructure\Text\Exceptions\TextRenderingFailedException;
 use ArtARTs36\MergeRequestLinter\Shared\DataStructure\Map;
 
 class MessageCreator
@@ -16,6 +17,7 @@ class MessageCreator
 
     /**
      * @param Map<string, mixed> $data
+     * @throws TextRenderingFailedException
      */
     public function create(string $template, Map $data): Message
     {