+++ src/Infrastructure/Text/Exceptions/TextRenderingFailedException.php
@@ -1,0 +1,10 @@
+<?php
+
+namespace ArtARTs36\MergeRequestLinter\Infrastructure\Text\Exceptions;
+
+use ArtARTs36\MergeRequestLinter\Shared\Exceptions\MergeRequestLinterException;
+
+class TextRenderingFailedException extends MergeRequestLinterException
+{
+    //
+}