+++ src/Presentation/Console/Exceptions/ApplicationNotCreatedException.php
@@ -1,0 +1,10 @@
+<?php
+
+namespace ArtARTs36\MergeRequestLinter\Presentation\Console\Exceptions;
+
+use ArtARTs36\MergeRequestLinter\Shared\Exceptions\MergeRequestLinterException;
+
+final class ApplicationNotCreatedException extends MergeRequestLinterException
+{
+    //
+}